# Conflicts:
#	Scripts/dbo.Trades.sql

+++ Scripts/20240507-datetime2-migration.sql
@@ -1,6 +1,10 @@
 -- dbo.Trades table update
 
-DROP INDEX IX_Trades_Id_AccountId_AssetPairId_OrderCreatedDate ON [dbo].[Trades];
+IF EXISTS (SELECT 'X'
+           FROM sys.indexes
+           WHERE name = 'IX_Trades_Id_AccountId_AssetPairId_OrderCreatedDate'
+             AND object_id = OBJECT_ID('dbo.Trades'))
+    DROP INDEX IX_Trades_Id_AccountId_AssetPairId_OrderCreatedDate ON [dbo].[Trades];
 
 ALTER TABLE [dbo].[Trades]
     ALTER COLUMN OrderCreatedDate datetime2 NOT NULL;
@@ -8,33 +12,52 @@ ALTER TABLE [dbo].[Trades]
 CREATE UNIQUE INDEX IX_Trades_Id_AccountId_AssetPairId_OrderCreatedDate
     ON [dbo].[Trades] (Id, AccountId, AssetPairId, OrderCreatedDate);
 
-DROP INDEX IX_Trades on [dbo].[Trades];
-DROP INDEX IX_Trades_Id_AccountId_AssetPairId_TradeTimestamp_Volume on [dbo].[Trades];
+IF EXISTS (SELECT 'X'
+           FROM sys.indexes
+           WHERE name = 'IX_Trades'
+             AND object_id = OBJECT_ID('dbo.Trades'))
+    DROP INDEX IX_Trades ON [dbo].[Trades];
+    
+IF EXISTS (SELECT 'X'
+           FROM sys.indexes
+           WHERE name = 'IX_Trades_Id_AccountId_AssetPairId_TradeTimestamp_Volume'
+             AND object_id = OBJECT_ID('dbo.Trades'))
+    DROP INDEX IX_Trades_Id_AccountId_AssetPairId_TradeTimestamp_Volume on [dbo].[Trades];
 
 ALTER TABLE [dbo].[Trades]
     ALTER COLUMN TradeTimestamp datetime2 NOT NULL;
 
 CREATE INDEX IX_Trades
-    ON dbo.Trades (TradeTimestamp);
+    ON [dbo].[Trades] (TradeTimestamp);
 
 CREATE UNIQUE INDEX IX_Trades_Id_AccountId_AssetPairId_TradeTimestamp_Volume
     ON [dbo].[Trades] (Id, AccountId, AssetPairId, TradeTimestamp, Volume);
 
 -- dbo.PositionsHistory table update
 
+/*
+- These indexes are only used in Lykke dev environment
 DROP INDEX IX_PositionsHistory_AccountId_OpenDate_CloseDate_15A9F on [dbo].[PositionsHistory]
 DROP INDEX IX_PositionsHistory_AccountId_HistoryTimestamp_2DFD0 on [dbo].[PositionsHistory]
-DROP INDEX IX_PositionsHistory_AccountId_OpenDate_CloseDate on [dbo].[PositionsHistory]
+ */
+
+IF EXISTS (SELECT 'X'
+           FROM sys.indexes
+           WHERE name = 'IX_PositionsHistory_AccountId_OpenDate_CloseDate'
+             AND object_id = OBJECT_ID('dbo.PositionsHistory'))
+    DROP INDEX IX_PositionsHistory_AccountId_OpenDate_CloseDate on [dbo].[PositionsHistory]
 
 ALTER TABLE [dbo].[PositionsHistory]
     ALTER COLUMN OpenDate datetime2 NOT NULL;
 
-create index IX_PositionsHistory_AccountId_OpenDate_CloseDate_15A9F
-    on dbo.PositionsHistory (AccountId, OpenDate, CloseDate) include (Id, AssetPairId, Direction, Volume, OpenTradeId,
+/*
+- These indexes are only used in Lykke dev environment
+CREATE INDEX IX_PositionsHistory_AccountId_OpenDate_CloseDate_15A9F
+    ON dbo.PositionsHistory (AccountId, OpenDate, CloseDate) INCLUDE (Id, AssetPairId, Direction, Volume, OpenTradeId,
                                                                       OpenPrice, OpenFxPrice, ClosePrice, CloseFxPrice)
 
-create index IX_PositionsHistory_AccountId_HistoryTimestamp_2DFD0
-    on dbo.PositionsHistory (AccountId, HistoryTimestamp) include (Id, DealId, Code, AssetPairId, Direction, Volume,
+CREATE INDEX IX_PositionsHistory_AccountId_HistoryTimestamp_2DFD0
+    ON dbo.PositionsHistory (AccountId, HistoryTimestamp) INCLUDE (Id, DealId, Code, AssetPairId, Direction, Volume,
                                                                    TradingConditionId, AccountAssetId,
                                                                    ExpectedOpenPrice, OpenMatchingEngineId, OpenDate,
                                                                    OpenTradeId, OpenPrice, OpenFxPrice, EquivalentAsset,
@@ -50,6 +73,7 @@ create index IX_PositionsHistory_AccountId_HistoryTimestamp_2DFD0
                                                                    OpenOrderVolume, FxAssetPairId,
                                                                    FxToAssetPairDirection, AdditionalInfo, ForceOpen,
                                                                    CorrelationId)
+ */
 
-create index IX_PositionsHistory_AccountId_OpenDate_CloseDate
-    on dbo.PositionsHistory (AccountId, OpenDate, CloseDate)
+CREATE INDEX IX_PositionsHistory_AccountId_OpenDate_CloseDate
+    ON [dbo].[PositionsHistory] (AccountId, OpenDate, CloseDate)
